Rework of damage detection
Adding two new layers "player damage", "enemy damage" which contains colliders that detect when an entity is hit. The damage colliders and colliders used for physics of rigid bodies are now separated. Lastly, the Bolt has been reworked so that is can accept some parameters before it is put into the world. And it does not need to lookup any components anymore (except when hitting something)