climbing proposal wip
most of the functionality seems to work here. Code still needs clean up though. The trick is a new handling of platforms. Each entity capable of using platforms has a platformer component which scans for nearby platforms and manager collision when the thing stands on top of it or if it walks through it.  The collisions are managed completely separately for every entity, no matter the layer of the entity. To scan for the platforms it uses layer DumbPlatform (because platform don t know anything about their state, they are manipulated by the Platformer) There are some small details that should be figured out. Such as whether the platformer should spy on its parrent to determine when it should ignore platforms or if the parrent should tell the platformer to ignore them. Also some refactoring to the climabilityHandler to be more generic (just using an interface). As part of polishing, I will move some files around to have a little more organization :) Lastly, I would propose to have Climability not care at all about whether it is being carried. Let it handle just climbing and its platform. Then we can make a component which is composed of Climability and Carriability and disable climability when it is being carried :)