Issue#45 platform proposal (#47)
* Basis of interaction system

Can pick up and drop objects by holding the F button

* Higher flexibility for carrying and adding of rigid bodies to props

* Creation of specific boss arena death states for enemies

* Implementation of alternative death states for enemies together with climbing mechanic

* Adding of climbing animations, fixing of climbing movement

* Filtering of interaction with interactables

* Fall damage fix

* Implementation of enabling platform when entity is at the top of the stack

* Getting rid of argument for isClimable method

* Fixing of aesthetics for different enemy sprites, enabling of climability by boolean

* Final touches

-implementation of ClimableComponent for enemies
-fixing of players animations

* Putting of climability into component

A bit of automatization, so there isn't that much stuff that has to be done later

* Fix of unfreezing rotation on spawning

* Fixes according to the feedback

* small adjustment in climability

* Implementation of cariability/climability into components

* Freezing position of IAmTop object, optimization of raycasts on platforms

* Drawings

* Platformer component

* climbing proposal wip

most of the functionality seems to work here. Code still needs clean up though. The trick is a new handling of platforms. Each entity capable of using platforms has a platformer component which scans for nearby platforms and manager collision when the thing stands on top of it or if it walks through it.  The collisions are managed completely separately for every entity, no matter the layer of the entity. To scan for the platforms it uses layer DumbPlatform (because platform don t know anything about their state, they are manipulated by the Platformer) There are some small details that should be figured out. Such as whether the platformer should spy on its parrent to determine when it should ignore platforms or if the parrent should tell the platformer to ignore them. Also some refactoring to the climabilityHandler to be more generic (just using an interface). As part of polishing, I will move some files around to have a little more organization :) Lastly, I would propose to have Climability not care at all about whether it is being carried. Let it handle just climbing and its platform. Then we can make a component which is composed of Climability and Carriability and disable climability when it is being carried :)

* climb down fix

* file organization

* stackability

* cleaning the object references

* fixup

* Rework of all carrible platforms

* finishing platforms, adjusting objects that use them. Cleaning up the warnings in the code

Co-authored-by: Jakub_Platzek <jakub.platzek@gmail.com>
Co-authored-by: Vlado <vlr@lindcapital.com>